Perfected update capacity page
Made the GUI look gorgeous - added drop down menus for options to make it more user-friendly.

Made it so the user can choose to update all days and all time slots, a specific day and a specific time slot, all days and a specific time slot, and a specific day and all time slots.

Added error checking for it as well.

WE ARE DONE! BOOM.